Explicitly poll until number of tasks is correct
`elbv2.get_waiter('target_in_service')` is not sufficient for waiting
until the correct number of tasks are running, only that there are some
tasks running. Instead we use our own `while` loop to wait until the
tasks have finished booting, or raise an exception if the timeout is
reached.